+++ orchestrator/api/api_v1/endpoints/settings.py
@@ -15,6 +15,7 @@ from http import HTTPStatus
 from typing import Optional
 
 from aiocache import Cache
+from fastapi import Query, WebSocket
 from fastapi.param_functions import Depends
 from fastapi.routing import APIRouter
 from oauth2_lib.fastapi import OIDCUserModel
@@ -22,11 +23,18 @@ from starlette.background import BackgroundTasks
 
 from orchestrator.api.error_handling import raise_status
 from orchestrator.db import EngineSettingsTable
-from orchestrator.schemas import EngineSettingsBaseSchema, EngineSettingsSchema, GlobalStatusEnum
+from orchestrator.schemas import (
+    EngineSettingsBaseSchema,
+    EngineSettingsSchema,
+    EngineSettingsSchemaResult,
+    GlobalStatusEnum,
+)
 from orchestrator.security import oidc_user
 from orchestrator.services import settings
 from orchestrator.services.processes import SYSTEM_USER
 from orchestrator.settings import app_settings
+from orchestrator.utils.json import json_dumps
+from orchestrator.websocket import WS_CHANNELS, websocket_enabled, websocket_manager
 
 router = APIRouter()
 
@@ -40,7 +48,7 @@ async def clear_cache(name: str, background_tasks: BackgroundTasks) -> None:
         background_tasks.add_task(cache.delete, f"/{name}*")
 
 
-@router.put("/status", response_model=EngineSettingsSchema)
+@router.put("/status", response_model=EngineSettingsSchemaResult)
 def set_global_status(
     body: EngineSettingsBaseSchema, user: Optional[OIDCUserModel] = Depends(oidc_user)
 ) -> EngineSettingsSchema:
@@ -67,10 +75,17 @@ def set_global_status(
         user_name = user.user_name if user else SYSTEM_USER
         settings.post_update_to_slack(EngineSettingsSchema.from_orm(result), user_name)
 
-    return generate_engine_status_response(result)
+    status_response = generate_engine_status_response(result)
+    if websocket_manager.enabled:
+        # send engine status to socket.
+        websocket_manager.broadcast_data(
+            [WS_CHANNELS.ENGINE_SETTINGS], {"engine-status": generate_engine_status_response(result)}
+        )
+
+    return status_response
 
 
-@router.get("/status", response_model=EngineSettingsSchema)
+@router.get("/status", response_model=EngineSettingsSchemaResult)
 def get_global_status() -> EngineSettingsSchema:
     """
     Retrieve the global status object.
@@ -81,6 +96,24 @@ def get_global_status() -> EngineSettingsSchema:
     """
     engine_settings = EngineSettingsTable.query.one()
     return generate_engine_status_response(engine_settings)
+
+
+@router.websocket("/ws-status/")
+@websocket_enabled
+async def websocket_get_global_status(websocket: WebSocket, token: str = Query(...)) -> None:
+    error = await websocket_manager.authorize(websocket, token)
+
+    await websocket.accept()
+    if error:
+        await websocket_manager.disconnect(websocket, reason=error)
+        return
+
+    engine_settings = EngineSettingsTable.query.one()
+
+    await websocket.send_text(json_dumps({"engine-status": generate_engine_status_response(engine_settings)}))
+
+    channel = WS_CHANNELS.ENGINE_SETTINGS
+    await websocket_manager.connect(websocket, channel)
 
 
 def generate_engine_status_response(engine_settings: EngineSettingsTable) -> EngineSettingsSchema: