+++ orchestrator/schemas/engine_settings.py
@@ -33,3 +33,8 @@ class EngineSettingsSchema(EngineSettingsBaseSchema):
 
     class Config:
         orm_mode = True
+
+
+class EngineSettingsSchemaResult(EngineSettingsSchema):
+    class Config:
+        orm_mode = False