+++ orchestrator/types.py
@@ -55,26 +55,33 @@ class SubscriptionLifecycle(strEnum):
 # The key is the Parent subscription life cycle status. The keys are lists of safe transitions for child subscriptions.
 SAFE_PARENT_TRANSITIONS_FOR_STATUS = {
     SubscriptionLifecycle.INITIAL: [
-        SubscriptionLifecycle.ACTIVE,
+        SubscriptionLifecycle.INITIAL,
+    ],
+    SubscriptionLifecycle.ACTIVE: [
+        SubscriptionLifecycle.INITIAL,
         SubscriptionLifecycle.PROVISIONING,
         SubscriptionLifecycle.MIGRATING,
+        SubscriptionLifecycle.ACTIVE,
+        SubscriptionLifecycle.TERMINATED,
+        SubscriptionLifecycle.DISABLED,
+    ],
+    SubscriptionLifecycle.MIGRATING: [
+        SubscriptionLifecycle.INITIAL,
+        SubscriptionLifecycle.MIGRATING,
+        SubscriptionLifecycle.TERMINATED,
+    ],
+    SubscriptionLifecycle.PROVISIONING: [
+        SubscriptionLifecycle.INITIAL,
+        SubscriptionLifecycle.PROVISIONING,
+        SubscriptionLifecycle.ACTIVE,
+        SubscriptionLifecycle.TERMINATED,
+    ],
+    SubscriptionLifecycle.TERMINATED: [SubscriptionLifecycle.INITIAL, SubscriptionLifecycle.TERMINATED],
+    SubscriptionLifecycle.DISABLED: [
+        SubscriptionLifecycle.INITIAL,
         SubscriptionLifecycle.DISABLED,
         SubscriptionLifecycle.TERMINATED,
     ],
-    SubscriptionLifecycle.ACTIVE: [
-        SubscriptionLifecycle.ACTIVE,
-        SubscriptionLifecycle.MIGRATING,
-    ],
-    SubscriptionLifecycle.MIGRATING: [SubscriptionLifecycle.ACTIVE],
-    SubscriptionLifecycle.PROVISIONING: [SubscriptionLifecycle.ACTIVE],
-    SubscriptionLifecycle.TERMINATED: [
-        SubscriptionLifecycle.INITIAL,
-        SubscriptionLifecycle.ACTIVE,
-        SubscriptionLifecycle.MIGRATING,
-        SubscriptionLifecycle.DISABLED,
-        SubscriptionLifecycle.PROVISIONING,
-    ],
-    SubscriptionLifecycle.DISABLED: [SubscriptionLifecycle.ACTIVE, SubscriptionLifecycle.TERMINATED],
 }