+++ orchestrator/websocket/__init__.py
@@ -12,6 +12,7 @@
 # limitations under the License.
 
 from asyncio import new_event_loop
+from functools import wraps
 from typing import Any, Dict, Optional, cast
 from urllib.parse import urlparse
 from uuid import UUID
@@ -22,6 +23,7 @@ from orchestrator.db import ProcessStepTable, ProcessTable
 from orchestrator.forms import generate_form
 from orchestrator.settings import AppSettings, app_settings
 from orchestrator.types import InputFormGenerator
+from orchestrator.utils.show_process import show_process
 from orchestrator.websocket.websocket_manager import WebSocketManager
 from orchestrator.workflow import ProcessStatus
 
@@ -29,6 +31,19 @@ logger = get_logger(__name__)
 
 
 broadcaster_type = urlparse(app_settings.WEBSOCKET_BROADCASTER_URL).scheme
+
+
+class WS_CHANNELS:
+    ALL_PROCESSES = "processes"
+    ENGINE_SETTINGS = "engine-settings"
+
+    @staticmethod
+    def SINGLE_PROCESS(pid: UUID) -> str:
+        return f"process_detail:{pid}"
+
+
+async def empty_fn(*args: tuple, **kwargs: Dict[str, Any]) -> None:
+    return
 
 
 class WrappedWebSocketManager:
@@ -47,6 +62,9 @@ class WrappedWebSocketManager:
             raise RuntimeWarning(
                 "No WebSocketManager configured at this time. Please pass WebSocketManager configuration to OrchestratorCore base_settings"
             )
+        if attr != "enabled" and not self.wrapped_websocket_manager.enabled:
+            logger.warning("Websockets are disabled, unable to access class methods")
+            return empty_fn
 
         return getattr(self.wrapped_websocket_manager, attr)
 
@@ -58,7 +76,7 @@ websocket_manager = cast(WebSocketManager, wrapped_websocket_manager)
 
 # The Global WebSocketManager is set after calling this function
 def init_websocket_manager(settings: AppSettings) -> WebSocketManager:
-    wrapped_websocket_manager.update(WebSocketManager(settings.WEBSOCKET_BROADCASTER_URL))
+    wrapped_websocket_manager.update(WebSocketManager(settings.ENABLE_WEBSOCKETS, settings.WEBSOCKET_BROADCASTER_URL))
     return websocket_manager
 
 
@@ -70,16 +88,13 @@ def create_process_step_websocket_data(
         form = generate_form(step_form, step.state, [])
 
     return {
-        "process": {
-            "assignee": process.assignee,
-            "step": process.last_step,
-            "status": process.last_status,
-            "last_modified": process.last_modified_at,
-        },
+        "process": show_process(process),
         "step": {
             "name": step.name,
+            "executed": int(step.executed_at.timestamp()),
             "status": step.status,
             "state": step.state,
+            "commit_hash": step.commit_hash,
             "form": form,
         },
     }
@@ -90,23 +105,42 @@ def is_process_active(p: Dict) -> bool:
 
 
 def send_process_step_data_to_websocket(pid: UUID, data: Dict) -> None:
-    channel = f"step_process:{pid}"
+    channel = WS_CHANNELS.SINGLE_PROCESS(pid)
 
     if not is_process_active(data["process"]):
         data["close"] = True
 
     loop = new_event_loop()
-    loop.run_until_complete(websocket_manager.broadcast_data(channel, data))
+    channels = [channel, WS_CHANNELS.ALL_PROCESSES, WS_CHANNELS.ENGINE_SETTINGS]
+    loop.run_until_complete(websocket_manager.broadcast_data(channels, data))
     try:
         loop.close()
     except Exception:  # noqa: S110
         pass
 
 
+async def empty_handler() -> None:
+    return
+
+
+def websocket_enabled(handler: Any) -> Any:
+    @wraps(handler)
+    @wraps(empty_handler)
+    async def wrapper(*args: tuple, **kwargs: Dict[str, Any]) -> Any:
+        if websocket_manager.enabled:
+            return await handler(*args, **kwargs)
+        else:
+            return await empty_handler()
+
+    return wrapper
+
+
 __all__ = [
     "websocket_manager",
     "init_websocket_manager",
-    "create_websocket_data",
-    "send_process_step_data_to_websocket",
+    "create_process_step_websocket_data",
     "is_process_active",
+    "send_process_step_data_to_websocket",
+    "websocket_enabled",
+    "WS_CHANNELS",
 ]