+++ orchestrator/websocket/managers/memory_websocket_manager.py
@@ -1,3 +1,16 @@
+# Copyright 2019-2020 SURF.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from typing import Dict, List, Optional, Union
 
 from fastapi import WebSocket, WebSocketDisconnect, status
@@ -43,14 +56,15 @@ class MemoryWebsocketManager:
             if len(self.connections_by_pid[channel]):
                 del self.connections_by_pid[channel]
 
-    async def broadcast_data(self, channel: str, data: Dict) -> None:
+    async def broadcast_data(self, channels: List[str], data: Dict) -> None:
         try:
-            if channel in self.connections_by_pid:
-                for connection in self.connections_by_pid[channel]:
-                    await connection.send_text(json_dumps(data))
+            for channel in channels:
+                if channel in self.connections_by_pid:
+                    for connection in self.connections_by_pid[channel]:
+                        await connection.send_text(json_dumps(data))
 
-            if "close" in data and data["close"]:
-                await self.disconnect_all(channel)
+                if "close" in data and data["close"]:
+                    await self.disconnect_all(channel)
         except (RuntimeError, ValueError):
             pass