+++ test/unit_tests/conftest.py
@@ -20,7 +20,6 @@ from orchestrator.db import ProductBlockTable, ProductTable, ResourceTypeTable, 
 from orchestrator.db.database import ENGINE_ARGUMENTS, SESSION_ARGUMENTS, BaseModel, Database, SearchQuery
 from orchestrator.domain import SUBSCRIPTION_MODEL_REGISTRY, SubscriptionModel
 from orchestrator.domain.base import ProductBlockModel
-from orchestrator.domain.lifecycle import change_lifecycle
 from orchestrator.forms import FormPage
 from orchestrator.services.translations import generate_translations
 from orchestrator.settings import app_settings
@@ -80,7 +79,10 @@ def db_uri(worker_id):
         # pytest is being run without any workers
         return database_uri
     url = make_url(database_uri)
-    url.database = f"{url.database}-{worker_id}"
+    if hasattr(url, "set"):
+        url = url.set(database=f"{url.database}-{worker_id}")
+    else:
+        url.database = f"{url.database}-{worker_id}"
     return str(url)
 
 
@@ -97,7 +99,10 @@ def database(db_uri):
     db.update(Database(db_uri))
     url = make_url(db_uri)
     db_to_create = url.database
-    url.database = "postgres"
+    if hasattr(url, "set"):
+        url = url.set(database="postgres")
+    else:
+        url.database = "postgres"
     engine = create_engine(url)
     with closing(engine.connect()) as conn:
         conn.execute("COMMIT;")
@@ -142,7 +147,8 @@ def db_session(database):
         try:
             yield
         finally:
-            trans.rollback()
+            if not trans._deactivated_from_connection:
+                trans.rollback()
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -446,10 +452,10 @@ def generic_subscription_1(generic_product_1, generic_product_type_1):
     gen_subscription.pb_1.rt_1 = "Value1"
     gen_subscription.pb_2.rt_2 = 42
     gen_subscription.pb_2.rt_3 = "Value2"
-    gen_subscription = change_lifecycle(gen_subscription, SubscriptionLifecycle.ACTIVE)
+    gen_subscription = SubscriptionModel.from_other_lifecycle(gen_subscription, SubscriptionLifecycle.ACTIVE)
     gen_subscription.description = "Generic Subscription One"
     gen_subscription.save()
-
+    db.session.commit()
     return str(gen_subscription.subscription_id)
 
 
@@ -460,9 +466,10 @@ def generic_subscription_2(generic_product_2, generic_product_type_2):
         generic_product_2.product_id, customer_id=CUSTOMER_ID, insync=True
     )
     gen_subscription.pb_3.rt_2 = 42
-    gen_subscription = change_lifecycle(gen_subscription, SubscriptionLifecycle.ACTIVE)
+    gen_subscription = SubscriptionModel.from_other_lifecycle(gen_subscription, SubscriptionLifecycle.ACTIVE)
     gen_subscription.description = "Generic Subscription One"
     gen_subscription.save()
+    db.session.commit()
 
     return str(gen_subscription.subscription_id)