+++ test/unit_tests/forms/test_generic_validators.py
@@ -11,6 +11,7 @@ from orchestrator.forms.validators import (
     Choice,
     ContactPersonList,
     DisplaySubscription,
+    Divider,
     Label,
     ListOfOne,
     ListOfTwo,
@@ -54,7 +55,9 @@ def test_constrained_list_constraints():
 
     with pytest.raises(ValidationError) as exc_info:
         UniqueConListModel(v=[1, 1, 1])
-    assert exc_info.value.errors() == [{"loc": ("v",), "msg": "Items must be unique", "type": "value_error"}]
+    assert exc_info.value.errors() == [
+        {"loc": ("v",), "msg": "the list has duplicated items", "type": "value_error.list.unique_items"}
+    ]
 
     with pytest.raises(ValidationError) as exc_info:
         UniqueConListModel(v=1)
@@ -89,7 +92,9 @@ def test_constrained_list_inherit_constraints():
 
     with pytest.raises(ValidationError) as exc_info:
         UniqueConListModel(v=[1, 1, 1])
-    assert exc_info.value.errors() == [{"loc": ("v",), "msg": "Items must be unique", "type": "value_error"}]
+    assert exc_info.value.errors() == [
+        {"loc": ("v",), "msg": "the list has duplicated items", "type": "value_error.list.unique_items"}
+    ]
 
     with pytest.raises(ValidationError) as exc_info:
         UniqueConListModel(v=1)
@@ -430,7 +435,9 @@ def test_choice_list_constraints():
 
     with pytest.raises(ValidationError) as exc_info:
         Form(choice=[1, 1, 1])
-    assert exc_info.value.errors() == [{"loc": ("choice",), "msg": "Items must be unique", "type": "value_error"}]
+    assert exc_info.value.errors() == [
+        {"loc": ("choice",), "msg": "the list has duplicated items", "type": "value_error.list.unique_items"}
+    ]
 
     with pytest.raises(ValidationError) as exc_info:
         Form(choice=1)
@@ -589,6 +596,18 @@ def test_display():
         "display_sub": None,
         "label": None,
         "migration_summary": None,
+    }
+
+
+def test_labels_with_value_and_dividers():
+    class Form(FormPage):
+        label: Label = "value"
+        divider: Divider
+
+    assert Form().dict() == {"label": "value", "divider": None}
+    assert Form(label="fob", divider="baz").dict() == {
+        "label": "value",
+        "divider": None,
     }