+++ CHANGELOG.md
@@ -7,7 +7,29 @@ reference the gitlab/github issue that is related to the change.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.0] - 2021-10-25
+
+### Breaking changes
+- Added explicit `subscription_id` parameter to the ProductBlockModel.
+- You must add a `product_block_relation` entry to the database if you need to stack product blocks in a domain model
+- Some changes in helper functions
+
+### Features
+- Introduced hierarchical relationships across subscription boundaries
+- Added some documentation
+- Added new form field that gives the ability to add a divider to forms
+- Better handling of empty optional fields
+- Added websocket failed task banner.
+
+## [0.0.21] - 2021-09-22
+- Improved docs
+- Allow pip -e installable packages.
+
+## [0.0.20] - 2021-08-05
+
+### Bugfix
+- fix bug in saving subscriptions, we now explicitly refresh mapped objects.
+
 
 ## [0.0.19] - 2021-07-19